chore: refactor ControlMenu component into smaller components
